[FLINK-21439][runtime] Move FailureResult into separate class
